Finished extended process class
Moved over implementations required for Round Robin from the Process class to the ExtendedProcess class, as they are not needded for a regular process.
